+++ src/main/kotlin/cloud/glitchdev/rfu/utils/Title.kt
@@ -1,13 +1,12 @@
 package cloud.glitchdev.rfu.utils
 
+import cloud.glitchdev.rfu.RiccioFishingUtils.mc
 import kotlinx.coroutines.*
-import net.minecraft.client.Minecraft
 import net.minecraft.network.chat.Component
 
 object Title {
-    private val queue: MutableList<TitleObj> = ArrayDeque()
+    private val queue: ArrayDeque<TitleObj> = ArrayDeque()
     private var isRunning = false
-    private val mc = Minecraft.getInstance() // Assuming you reference this
 
     fun showTitle(title: String, subTitle: String = "", fadeIn: Int = 5, duration: Int = 10, fadeOut: Int = 5, condition: () -> Boolean = { true }) {
         queue.add(TitleObj(title, subTitle, fadeIn, duration, fadeOut, condition))