+++ CHANGELOG.md
@@ -23,3 +23,4 @@
 ### Other:
 - Improved description for Fishing Downtime Limit
 - Made alerts show up for less time
+- Added a queueing system for party messages to avoid sending them too fast